+++ CHANGELOG.md
@@ -25,7 +25,8 @@ N/A
 
 ### Changed
 
-Make input event immutable to PostOffice's send, asyncRequest and request API
+1. Make input event immutable to PostOffice's send and request API
+2. Consistent temporary stream folder name for Java and Node.js under /tmp/composable
 
 ---
 ## Version 4.2.4, 1/30/2025